+++ Dfe.Academies.External.Web/AcademiesAPIResponseModels/Schools/SchoolsSearchDto.cs
@@ -1,0 +1,35 @@
+﻿using System.Text.Json.Serialization;
+
+namespace Dfe.Academies.External.Web.AcademiesAPIResponseModels.Schools;
+
+/// <summary>
+/// To de-serialize response from :-
+/// {{api-host}}/establishments?Name=wise&api-version=V1&Urn=101934&ukprn=10006563
+/// </summary>
+public sealed class SchoolsSearchDto
+{
+	[JsonPropertyName("urn")]
+	public int Urn { get; }
+
+	[JsonPropertyName("name")]
+	public string Name { get; }
+
+	[JsonPropertyName("ukprn")]
+	public string Ukprn { get; }
+
+	// TODO MR:- current establishments search API doesn't return address details
+	[JsonPropertyName("street")]
+	public string Street { get; }
+
+	[JsonPropertyName("town")]
+	public string Town { get; }
+
+	[JsonPropertyName("FullUkPostcode")]
+	public string FullUkPostcode { get; }
+
+	[JsonConstructor]
+	public SchoolsSearchDto(string name, int urn, string ukprn,
+		string street, string town, string fullUkPostcode) =>
+		(Name, Urn, Ukprn, Street, Town, FullUkPostcode) =
+		(name, urn, ukprn, street, town, fullUkPostcode);
+}
